% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_specialKey.R
\name{extra_specialKey}
\alias{extra_specialKey}
\title{Send special key input}
\usage{
extra_specialKey(key = "ESC", failSafe = TRUE)
}
\arguments{
\item{key}{"ESC", "ENTER", "TAB", "CTRL", or "ALT" as string}

\item{failSafe}{Logical flag to allow for emergency stops. Default is TRUE}
}
\description{
Send special keys, identical to \code{\link{rMouse::specialKey()}} but with "CTRL" and "ALT" added
}
\examples{
extra_specialKey("ALT")
extra_specialKey("CTRL")

}
