% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hold_specialKey.R
\name{hold_specialKey}
\alias{hold_specialKey}
\title{Hold special key input}
\usage{
hold_specialKey(key = "CTRL", failSafe = TRUE)
}
\arguments{
\item{key}{"CTRL" or "ALT" as string}

\item{failSafe}{Logical flag to allow for emergency stops. Default is TRUE}
}
\description{
Takes special key input and holds it until \code{\link{release}} is called
}
\examples{
rMouse::type("text to copy"); hold_specialKey("CTRL"); rMouse::type("ac"); release_specialKey("CTRL")

}
