% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_specialKey.R
\name{release_specialKey}
\alias{release_specialKey}
\title{Release special key input}
\usage{
release_specialKey(key = "CTRL", failSafe = TRUE)
}
\arguments{
\item{key}{"CTRL" or "ALT" as string}

\item{failSafe}{Logical flag to allow for emergency stops. Default is TRUE}
}
\description{
Takes special key input and releases it
}
\examples{
rMouse::type("text to copy"); hold_specialKey("CTRL"); rMouse::type("ac"); release_specialKey("CTRL")

}
